+++ x/auth/tx/encode_decode_test.go
@@ -12,10 +12,7 @@ import (
 	"github.com/cosmos/cosmos-sdk/codec"
 	codectypes "github.com/cosmos/cosmos-sdk/codec/types"
 	"github.com/cosmos/cosmos-sdk/testutil/testdata"
-	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
 	"github.com/cosmos/cosmos-sdk/types/tx"
-	signingtypes "github.com/cosmos/cosmos-sdk/types/tx/signing"
-	"github.com/cosmos/cosmos-sdk/x/auth/signing"
 )
 
 func TestDefaultTxDecoderError(t *testing.T) {
@@ -45,11 +42,10 @@ func TestUnknownFields(t *testing.T) {
 	decoder := DefaultTxDecoder(cdc)
 
 	tests := []struct {
-		name           string
-		body           *testdata.TestUpdatedTxBody
-		authInfo       *testdata.TestUpdatedAuthInfo
-		shouldErr      bool
-		shouldAminoErr string
+		name      string
+		body      *testdata.TestUpdatedTxBody
+		authInfo  *testdata.TestUpdatedAuthInfo
+		shouldErr bool
 	}{
 		{
 			name: "no new fields should pass",
@@ -65,9 +61,8 @@ func TestUnknownFields(t *testing.T) {
 				Memo:                         "foo",
 				SomeNewFieldNonCriticalField: "blah",
 			},
-			authInfo:       &testdata.TestUpdatedAuthInfo{},
-			shouldErr:      false,
-			shouldAminoErr: fmt.Sprintf("%s: %s", aminoNonCriticalFieldsError, sdkerrors.ErrInvalidRequest.Error()),
+			authInfo:  &testdata.TestUpdatedAuthInfo{},
+			shouldErr: false,
 		},
 		{
 			name: "critical fields in TxBody should error on decode",
@@ -121,15 +116,6 @@ func TestUnknownFields(t *testing.T) {
 				require.Error(t, err)
 			} else {
 				require.NoError(t, err)
-			}
-
-			if tt.shouldAminoErr != "" {
-				handler := signModeLegacyAminoJSONHandler{}
-				decoder := DefaultTxDecoder(codec.NewProtoCodec(codectypes.NewInterfaceRegistry()))
-				theTx, err := decoder(txBz)
-				require.NoError(t, err)
-				_, err = handler.GetSignBytes(signingtypes.SignMode_SIGN_MODE_LEGACY_AMINO_JSON, signing.SignerData{}, theTx)
-				require.EqualError(t, err, tt.shouldAminoErr)
 			}
 		})
 	}